requirements.txt
pyOpenSSL
werkzeug
djangorestframework
django
pytest
pydantic2
PyJWT
django-extensions
psycopg2-binary
drf-yasg
coreapi
django-cors-headers
requests>=2.31.0
redis>=5.0.0
django-redis>=5.4.0
